refactor & rename & consolidate
