+++ src/python/tests/test_cluster.py
@@ -2,15 +2,7 @@ import os, csv
 import pytest
 
 from . import sourmash_tst_utils as utils
-
-def get_test_data(filename):
-    thisdir = os.path.dirname(__file__)
-    return os.path.join(thisdir, 'test-data', filename)
-
-def make_file_list(filename, paths):
-    with open(filename, 'wt') as fp:
-        fp.write("\n".join(paths))
-        fp.write("\n")
+from .sourmash_tst_utils import get_test_data, make_file_list
 
 
 def test_installed(runtmp):
