+++ src/python/tests/test_manysearch.py
@@ -4,17 +4,7 @@ import pandas
 import sourmash
 
 from . import sourmash_tst_utils as utils
-
-
-def get_test_data(filename):
-    thisdir = os.path.dirname(__file__)
-    return os.path.join(thisdir, 'test-data', filename)
-
-
-def make_file_list(filename, paths):
-    with open(filename, 'wt') as fp:
-        fp.write("\n".join(paths))
-        fp.write("\n")
+from .sourmash_tst_utils import (get_test_data, make_file_list, zip_siglist)
 
 
 def test_installed(runtmp):
@@ -23,10 +13,6 @@ def test_installed(runtmp):
 
     assert 'usage:  manysearch' in runtmp.last_result.err
 
-def zip_siglist(runtmp, siglist, db):
-    runtmp.sourmash('sig', 'cat', siglist,
-                    '-o', db)
-    return db
 
 def index_siglist(runtmp, siglist, db, ksize=31, scaled=1000, moltype='DNA'):
     # build index
@@ -35,8 +21,6 @@ def index_siglist(runtmp, siglist, db, ksize=31, scaled=1000, moltype='DNA'):
                     '--moltype', moltype)
     return db
 
-@pytest.mark.parametrize("zip_query", [False, True])
-@pytest.mark.parametrize("zip_against", [False, True])
 def test_simple(runtmp, zip_query, zip_against):
     # test basic execution!
     query_list = runtmp.output('query.txt')
@@ -192,7 +176,6 @@ def test_simple_abund(runtmp):
     assert total_weighted_hashes == 73489
 
 
-@pytest.mark.parametrize("zip_query", [False, True])
 def test_simple_indexed(runtmp, zip_query):
     # test basic execution!
     query_list = runtmp.output('query.txt')
@@ -249,8 +232,6 @@ def test_simple_indexed(runtmp, zip_query):
                 assert query_ani == 0.9772
 
 
-@pytest.mark.parametrize("indexed", [False, True])
-@pytest.mark.parametrize("zip_query", [False, True])
 def test_simple_with_cores(runtmp, capfd, indexed, zip_query):
     # test basic execution with -c argument (that it runs, at least!)
     query_list = runtmp.output('query.txt')
@@ -283,8 +264,6 @@ def test_simple_with_cores(runtmp, capfd, indexed, zip_query):
     assert " using 4 threads" in result.err
 
 
-@pytest.mark.parametrize("indexed", [False, True])
-@pytest.mark.parametrize("zip_query", [False, True])
 def test_simple_threshold(runtmp, indexed, zip_query):
     # test with a simple threshold => only 3 results
     query_list = runtmp.output('query.txt')
@@ -313,7 +292,6 @@ def test_simple_threshold(runtmp, indexed, zip_query):
     assert len(df) == 3
 
 
-@pytest.mark.parametrize("indexed", [False, True])
 def test_simple_manifest(runtmp, indexed):
     # test with a simple threshold => only 3 results
     query_list = runtmp.output('query.txt')
@@ -347,8 +325,6 @@ def test_simple_manifest(runtmp, indexed):
     assert len(df) == 3
 
 
-@pytest.mark.parametrize("indexed", [False, True])
-@pytest.mark.parametrize("zip_query", [False, True])
 def test_missing_query(runtmp, capfd, indexed, zip_query):
     # test with a missing query list
     query_list = runtmp.output('query.txt')
@@ -379,7 +355,6 @@ def test_missing_query(runtmp, capfd, indexed, zip_query):
     assert 'Error: No such file or directory' in captured.err
 
 
-@pytest.mark.parametrize("indexed", [False, True])
 def test_sig_query(runtmp, capfd, indexed):
     # test with a single sig query (a .sig.gz file)
     against_list = runtmp.output('against.txt')
@@ -399,7 +374,6 @@ def test_sig_query(runtmp, capfd, indexed):
                         '-o', output)
 
 
-@pytest.mark.parametrize("indexed", [False, True])
 def test_bad_query_2(runtmp, capfd, indexed):
     # test with a bad query list (a missing file)
     query_list = runtmp.output('query.txt')
@@ -453,7 +427,6 @@ def test_bad_query_3(runtmp, capfd):
     assert 'InvalidArchive' in captured.err
 
 
-@pytest.mark.parametrize("indexed", [False, True])
 def test_missing_against(runtmp, capfd, indexed):
     # test with a missing against list
     query_list = runtmp.output('query.txt')
@@ -524,7 +497,6 @@ def test_bad_against(runtmp, capfd):
     assert "WARNING: 1 search paths failed to load. See error messages above." in captured.err
 
 
-@pytest.mark.parametrize("indexed", [False, True])
 def test_empty_query(runtmp, indexed, capfd):
     # test with an empty query list
     query_list = runtmp.output('query.txt')
@@ -552,8 +524,6 @@ def test_empty_query(runtmp, indexed, capfd):
     assert "No query signatures loaded, exiting." in captured.err
 
 
-@pytest.mark.parametrize("indexed", [False, True])
-@pytest.mark.parametrize("zip_query", [False, True])
 def test_nomatch_query(runtmp, capfd, indexed, zip_query):
     # test a non-matching (diff ksize) in query; do we get warning message?
     query_list = runtmp.output('query.txt')
@@ -584,8 +554,6 @@ def test_nomatch_query(runtmp, capfd, indexed, zip_query):
     assert 'WARNING: skipped 1 query paths - no compatible signatures.' in captured.err
 
 
-@pytest.mark.parametrize("zip_against", [False, True])
-@pytest.mark.parametrize("indexed", [False, True])
 def test_load_only_one_bug(runtmp, capfd, indexed, zip_against):
     # check that we behave properly when presented with multiple against
     # sketches
@@ -619,8 +587,6 @@ def test_load_only_one_bug(runtmp, capfd, indexed, zip_against):
     assert not 'WARNING: no compatible sketches in path ' in captured.err
 
 
-@pytest.mark.parametrize("zip_query", [False, True])
-@pytest.mark.parametrize("indexed", [False, True])
 def test_load_only_one_bug_as_query(runtmp, capfd, indexed, zip_query):
     # check that we behave properly when presented with multiple query
     # sketches in one file, with only one matching.
@@ -656,8 +622,6 @@ def test_load_only_one_bug_as_query(runtmp, capfd, indexed, zip_query):
     assert not 'WARNING: no compatible sketches in path ' in captured.err
 
 
-@pytest.mark.parametrize("zip_query", [False, True])
-@pytest.mark.parametrize("indexed", [False, True])
 def test_md5(runtmp, indexed, zip_query):
     # test that md5s match what was in the original files, not downsampled etc.
     query_list = runtmp.output('query.txt')
