+++ src/python/tests/test_multisearch.py
@@ -5,17 +5,7 @@ import pandas
 import sourmash
 
 from . import sourmash_tst_utils as utils
-
-
-def get_test_data(filename):
-    thisdir = os.path.dirname(__file__)
-    return os.path.join(thisdir, 'test-data', filename)
-
-
-def make_file_list(filename, paths):
-    with open(filename, 'wt') as fp:
-        fp.write("\n".join(paths))
-        fp.write("\n")
+from .sourmash_tst_utils import (get_test_data, make_file_list, zip_siglist)
 
 
 def test_installed(runtmp):
@@ -24,13 +14,7 @@ def test_installed(runtmp):
 
     assert 'usage:  multisearch' in runtmp.last_result.err
 
-def zip_siglist(runtmp, siglist, db):
-    runtmp.sourmash('sig', 'cat', siglist,
-                    '-o', db)
-    return db
 
-@pytest.mark.parametrize("zip_query", [False, True])
-@pytest.mark.parametrize("zip_db", [False, True])
 def test_simple_no_ani(runtmp, zip_query, zip_db):
     # test basic execution!
     query_list = runtmp.output('query.txt')
@@ -99,8 +83,6 @@ def test_simple_no_ani(runtmp, zip_query, zip_db):
                 assert intersect_hashes == 2529
 
 
-@pytest.mark.parametrize("zip_query", [False, True])
-@pytest.mark.parametrize("zip_db", [False, True])
 def test_simple_ani(runtmp, zip_query, zip_db):
     # test basic execution!
     query_list = runtmp.output('query.txt')
@@ -186,8 +168,6 @@ def test_simple_ani(runtmp, zip_query, zip_db):
                 assert max_ani == 0.9772
 
 
-@pytest.mark.parametrize("zip_query", [False, True])
-@pytest.mark.parametrize("zip_db", [False, True])
 def test_simple_threshold(runtmp, zip_query, zip_db):
     # test with a simple threshold => only 3 results
     query_list = runtmp.output('query.txt')
@@ -243,7 +223,6 @@ def test_simple_manifest(runtmp):
     assert len(df) == 3
 
 
-@pytest.mark.parametrize("zip_query", [False, True])
 def test_missing_query(runtmp, capfd, zip_query):
     # test with a missing query list
     query_list = runtmp.output('query.txt')
@@ -344,7 +323,6 @@ def test_bad_query_3(runtmp, capfd):
     assert 'InvalidArchive' in captured.err
 
 
-@pytest.mark.parametrize("zip_db", [False, True])
 def test_missing_against(runtmp, capfd, zip_db):
     # test with a missing against list
     query_list = runtmp.output('query.txt')
@@ -445,7 +423,6 @@ def test_empty_query(runtmp, capfd):
     # @CTB
 
 
-@pytest.mark.parametrize("zip_query", [False, True])
 def test_nomatch_query(runtmp, capfd, zip_query):
     # test a non-matching (diff ksize) in query; do we get warning message?
     query_list = runtmp.output('query.txt')
@@ -474,7 +451,6 @@ def test_nomatch_query(runtmp, capfd, zip_query):
     assert 'WARNING: skipped 1 query paths - no compatible signatures' in captured.err
 
 
-@pytest.mark.parametrize("zip_db", [False, True])
 def test_load_only_one_bug(runtmp, capfd, zip_db):
     # check that we behave properly when presented with multiple against
     # sketches
@@ -506,7 +482,6 @@ def test_load_only_one_bug(runtmp, capfd, zip_db):
     assert not 'WARNING: no compatible sketches in path' in captured.err
 
 
-@pytest.mark.parametrize("zip_query", [False, True])
 def test_load_only_one_bug_as_query(runtmp, capfd, zip_query):
     # check that we behave properly when presented with multiple query
     # sketches in one file, with only one matching.
@@ -538,8 +513,6 @@ def test_load_only_one_bug_as_query(runtmp, capfd, zip_query):
     assert not 'WARNING: no compatible sketches in path ' in captured.err
 
 
-@pytest.mark.parametrize("zip_query", [False, True])
-@pytest.mark.parametrize("zip_db", [False, True])
 def test_md5(runtmp, zip_query, zip_db):
     # test that md5s match what was in the original files, not downsampled etc.
     query_list = runtmp.output('query.txt')
